requirements.txt
python-dotenv
aiohttp
Jinja2